requirements.txt
opencv-python==4.5.3.56
PyAutoGUI==0.9.52
mss==6.1.0
numpy==1.19.5
Pillow==8.3.1
pytorchyolo==1.5.1
